+++ dom/media/Benchmark.cpp
@@ -11,6 +11,7 @@
 #include "PDMFactory.h"
 #include "WebMDemuxer.h"
 #include "mozilla/Preferences.h"
+#include "mozilla/SharedThreadPool.h"
 #include "mozilla/dom/ContentChild.h"
 
 #ifndef MOZ_WIDGET_ANDROID
