+++ dom/media/CubebUtils.cpp
@@ -10,6 +10,8 @@
 #include "nsDebug.h"
 #include "nsString.h"
 #include "mozilla/Preferences.h"
+#include "mozilla/Services.h"
+#include "mozilla/Sprintf.h"
 #include "mozilla/StaticMutex.h"
 #include "mozilla/StaticPtr.h"
 #include "mozilla/Telemetry.h"
