+++ dom/media/DOMMediaStream.cpp
@@ -14,6 +14,7 @@
 #include "mozilla/dom/MediaStreamTrackEvent.h"
 #include "mozilla/dom/LocalMediaStreamBinding.h"
 #include "mozilla/dom/AudioNode.h"
+#include "AudioCaptureStream.h"
 #include "AudioChannelAgent.h"
 #include "mozilla/dom/AudioTrack.h"
 #include "mozilla/dom/AudioTrackList.h"
@@ -22,6 +23,7 @@
 #include "mozilla/dom/HTMLCanvasElement.h"
 #include "mozilla/media/MediaUtils.h"
 #include "MediaStreamGraph.h"
+#include "MediaStreamListener.h"
 #include "AudioStreamTrack.h"
 #include "VideoStreamTrack.h"
 #include "Layers.h"
