+++ dom/media/MediaStreamTrack.cpp
@@ -7,8 +7,10 @@
 
 #include "DOMMediaStream.h"
 #include "MediaStreamGraph.h"
+#include "nsContentUtils.h"
 #include "nsIUUIDGenerator.h"
 #include "nsServiceManagerUtils.h"
+#include "MediaStreamError.h"
 #include "MediaStreamListener.h"
 #include "systemservices/MediaUtils.h"
 
@@ -20,6 +22,8 @@
 
 static mozilla::LazyLogModule gMediaStreamTrackLog("MediaStreamTrack");
 #define LOG(type, msg) MOZ_LOG(gMediaStreamTrackLog, type, msg)
+
+using namespace mozilla::media;
 
 namespace mozilla {
 namespace dom {
