+++ dom/media/gmp/GMPDecryptorParent.cpp
@@ -4,7 +4,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "GMPDecryptorParent.h"
+
 #include "GMPContentParent.h"
+#include "GMPUtils.h"
 #include "MediaData.h"
 #include "mozilla/Unused.h"
