+++ dom/media/gmp/GMPUtils.cpp
@@ -10,8 +10,10 @@
 #include "nsCOMPtr.h"
 #include "nsLiteralString.h"
 #include "nsCRTGlue.h"
+#include "mozIGeckoMediaPluginService.h"
 #include "mozilla/Base64.h"
 #include "nsISimpleEnumerator.h"
+#include "prio.h"
 
 namespace mozilla {
