+++ dom/media/imagecapture/CaptureTask.cpp
@@ -15,24 +15,6 @@
 
 namespace mozilla {
 
-class CaptureTask::MediaStreamEventListener : public MediaStreamTrackListener
-{
-public:
-  explicit MediaStreamEventListener(CaptureTask* aCaptureTask)
-    : mCaptureTask(aCaptureTask) {};
-
-  // MediaStreamListener methods.
-  void NotifyEnded() override
-  {
-    if(!mCaptureTask->mImageGrabbedOrTrackEnd) {
-      mCaptureTask->PostTrackEndEvent();
-    }
-  }
-
-private:
-  CaptureTask* mCaptureTask;
-};
-
 CaptureTask::CaptureTask(dom::ImageCapture* aImageCapture)
   : mImageCapture(aImageCapture)
   , mEventListener(new MediaStreamEventListener(this))
