+++ dom/media/imagecapture/moz.build
@@ -8,7 +8,7 @@ EXPORTS.mozilla.dom += [
     'ImageCapture.h'
 ]
 
-UNIFIED_SOURCES += [
+SOURCES += [
     'CaptureTask.cpp',
     'ImageCapture.cpp',
 ]
