+++ dom/media/moz.build
@@ -184,7 +184,7 @@ EXPORTS.mozilla.dom += [
     'VideoTrackList.h',
 ]
 
-UNIFIED_SOURCES += [
+SOURCES += [
     'AccurateSeekTask.cpp',
     'ADTSDecoder.cpp',
     'ADTSDemuxer.cpp',
@@ -201,6 +201,7 @@ UNIFIED_SOURCES += [
     'CanvasCaptureMediaStream.cpp',
     'CubebUtils.cpp',
     'DecoderDoctorDiagnostics.cpp',
+    'DecoderTraits.cpp',
     'DOMMediaStream.cpp',
     'EncodedBufferCache.cpp',
     'FileBlockCache.cpp',
@@ -255,11 +256,6 @@ UNIFIED_SOURCES += [
 
 if CONFIG['OS_TARGET'] == 'WINNT':
   SOURCES += [ 'ThreadPoolCOMListener.cpp' ]
-
-# DecoderTraits.cpp needs to be built separately because of Mac OS X headers.
-SOURCES += [
-    'DecoderTraits.cpp',
-]
 
 # Some codec-related code uses multi-character constants, which GCC and clang
 # warn about. Suppress turning this warning into an error.
