+++ dom/media/platforms/agnostic/eme/moz.build
@@ -11,7 +11,7 @@ EXPORTS += [
     'SamplesWaitingForKey.h',
 ]
 
-UNIFIED_SOURCES += [
+SOURCES += [
     'EMEAudioDecoder.cpp',
     'EMEDecoderModule.cpp',
     'EMEVideoDecoder.cpp',
