+++ dom/media/platforms/moz.build
@@ -16,7 +16,7 @@ EXPORTS += [
     'wrappers/H264Converter.h'
 ]
 
-UNIFIED_SOURCES += [
+SOURCES += [
     'agnostic/AgnosticDecoderModule.cpp',
     'agnostic/BlankDecoderModule.cpp',
     'agnostic/OpusDecoder.cpp',
@@ -42,7 +42,7 @@ if CONFIG['MOZ_WMF']:
 
 if CONFIG['MOZ_FFVPX'] or CONFIG['MOZ_FFMPEG']:
     # common code to either FFmpeg or FFVPX
-    UNIFIED_SOURCES += [
+    SOURCES += [
         'ffmpeg/FFmpegLibWrapper.cpp',
     ]
 
@@ -60,7 +60,7 @@ if CONFIG['MOZ_AV1']:
     EXPORTS += [
         'agnostic/AOMDecoder.h',
     ]
-    UNIFIED_SOURCES += [
+    SOURCES += [
         'agnostic/AOMDecoder.cpp',
     ]
 
@@ -68,7 +68,7 @@ if CONFIG['MOZ_APPLEMEDIA']:
   EXPORTS += [
       'apple/AppleDecoderModule.h',
   ]
-  UNIFIED_SOURCES += [
+  SOURCES += [
       'apple/AppleATDecoder.cpp',
       'apple/AppleCMLinker.cpp',
       'apple/AppleDecoderModule.cpp',
@@ -85,7 +85,7 @@ if CONFIG['MOZ_WIDGET_TOOLKIT'] == 'android':
     EXPORTS += [
         'android/AndroidDecoderModule.h',
     ]
-    UNIFIED_SOURCES += [
+    SOURCES += [
         'android/AndroidDecoderModule.cpp',
         'android/MediaCodecDataDecoder.cpp',
         'android/RemoteDataDecoder.cpp',
