+++ dom/media/platforms/omx/OmxPlatformLayer.cpp
@@ -6,6 +6,8 @@
 
 #include "OmxPlatformLayer.h"
 
+#include "OmxDataDecoder.h"
+
 #include "OMX_VideoExt.h" // For VP8.
 
 #include "VPXDecoder.h"
