+++ dom/media/platforms/wmf/moz.build
@@ -14,6 +14,7 @@ EXPORTS += [
     'WMFUtils.h',
     'WMFVideoMFTManager.h',
 ]
+
 UNIFIED_SOURCES += [
     'DXVA2Manager.cpp',
     'MFTDecoder.cpp',
